+++ verticapy/core/tablesample.py
@@ -14,11 +14,6 @@ OR CONDITIONS OF ANY KIND, either express or implied.
 See the  License for the specific  language governing
 permissions and limitations under the License.
 """
-
-#
-#
-# Modules
-#
 # Standard Python Modules
 import math, decimal, datetime, copy
 import numpy as np
